Conflicts:
	pom.xml

+++ core/broadleaf-framework-web/src/main/java/org/broadleafcommerce/core/web/api/wrapper/RelatedProductWrapper.java
@@ -1,28 +1,33 @@
 /*
- * Copyright 2008-2013 the original author or authors.
- *
+ * #%L
+ * BroadleafCommerce Framework Web
+ * %%
+ * Copyright (C) 2009 - 2013 Broadleaf Commerce
+ * %%
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
- *        http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ * #L%
  */
-
 package org.broadleafcommerce.core.web.api.wrapper;
 
-import org.broadleafcommerce.core.catalog.domain.RelatedProduct;
+import java.math.BigDecimal;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
+
+import org.broadleafcommerce.core.catalog.domain.RelatedProduct;
 
 /**
  * This is a JAXB wrapper for RelatedProducts
@@ -38,7 +43,7 @@ public class RelatedProductWrapper extends BaseWrapper implements APIWrapper<Rel
     protected Long id;
     
     @XmlElement
-    protected Long sequence;
+    protected BigDecimal sequence;
     
     @XmlElement
     protected String promotionalMessage;
